重写清理逻辑 还是觉得按照一个神经元一个线程写好一些

+++ cluo-brain/src/main/java/com/canfuu/cluo/brain/unit/HiddenUnitOutput.java
@@ -12,7 +12,7 @@ public class HiddenUnitOutput  extends CommonEntity {
 
 
     // 每次向后传递的value数量
-    private int transValue = 3;
+    private byte transValue = 3;
 
     public HiddenUnitOutput() {
     }
